Large beta update
Temporary thorsmex logo added to the header, a few of the items were changed, all item images changed for square versions, and large item detail changes, including the removal of UUID (not needed due to the way that the model number was implemented, model number acts as a unique id would). BIG styling changes for the header, and product pages.