Database implemented
The productModel was now completely replaced with the MongoDB database, that imports all of it's data from products.csv. To update the information in the database, products.csv needs to be updated. All other changes are fitting this database into the application, with some slight changes to what data is displayed on the product pages, as well as a 'product image not found' image that is displayed for when a product's image has not been found. I still need to import all of the product images, as well as adding more images for each product. Most images currently do not have an image until then.