Reorganization of files, housekeeping update
Reorganization of files and moved the "Scroll to top" button functionality to the app.js so it applies to every page on the site. Also added a component that is imported into App.js that automatically scrolls to the top of every page. Icon added for "Scroll to top" button.