More updates to catalog styling
More updates to the styling of the overall catalog, including moving all of the sort features to the sidebar. There were also some major housekeeping updates made, with the css greatly cleaned up as well as having catalog-container and catalog-categories div combined into catalog-container. The sidebar is now named simply "sidebar", which is inside of the "sidebar-wrapper"