New information added to products.csv and model
Finished adding and formatting the specs given by Sebastian from Thorsman. Also added in Price, dimensions, count and weight for boxes and pallet, and updated the current ones to have _indv at the end, meaning it is for only one instead of a box of products or a pallet of boxes.