"Extended Image" functionality extended
Now, when the extended image is shown, you can click "off" the image to close it. The close button functionality is still there and functional as well, but this gives the user another option to close the extended image. The extended image feature is now complete, besides some potential style changes that may possibly be made in the application's future.