Updated how data is pulled from the backend, backend server starting imports newest xlsx data
Updated the App.js fetchProducts function to check if the data imported from the backend server is different from the local storage, if it is different from the local storage, it will replace the local storage, if not, it will use the local storage. Also updated the backend server to always import the newest data from the xlsx each time the server is run.