Mobile version style updates
Adjusted the mobile version styling, now the grid of products on the mobile version of the catalog will have two columns instead of one. Also started on a new version of the product page mobile version, which still needs more work, this will be my next goal before I revamp the export as PDF features.