Set up Catalog and styling for selected item exporting functionality
Set up the catalog.js and styling in order for the selected products functionality to work properly. There will now be an invisible div "selected-product-grid", that all the selected products will be put into. When the export selected products as pdf function is called, it will temporarily make this div visible, export it as a pdf, then make it hidden again after it is done exporting it. This functionality is WIP, but the buttons, as well as the div and the styling for this to work properly have been added.