"Load More" Button fixed, logic behind exporting all as PDF made dynamic, small sidebar style update
The "load more" button bug has been fixed, it can now be used exactly as before the export as PDF updates were made. The logic for this export all as PDF function has been updated as well, making it more dynamic (it now sets the items per page dynamically, instead of just setting it to 200). This was not a problem yet, but in the future could become an issue if more products were added. Also made a small style change to the sidebar.