Big changes made to the export as PDF features
All of the "export as PDF" options now run through the same exportAsPDF function, this greatly helps with code readability as well with making the functionality simpler. Also, the margins should now correctly be dynamically calculated when the catalog is exported. There was also a "loading circle" added for when there is a PDF export being generated.