Select Categories buttons updated
Bug fix where clicking a category "page link" would change the visual state of a different "select" button. This is much closer to the final vision for this feature, the only issue now is that you cannot select all items in a category you are not currently able to view.