Small updates to product page and header
Small styling changes to the product page and header, mostly targeting more small style issues on the mobile page. The overlay buttons displaying over the expanded images shouldn't be over the images any more. The size of the header logo div does not go off the header any more. The header menu icon was moved slightly to the right.