Various fixes applied to email form submission process
The email form alert bug has been fixed, meaning the alerts from the form are working properly again. I also fixed the response from the server to properly show in console, and made it so that the form fields will be reset to blank after the form is successfully submitted.